changed float to double

+++ mandelbrot/kernel.cl
@@ -8,15 +8,15 @@
 #define MAX_ITERATIONS 200
 
 #define IMAGINARY_POS(y, height)\
-    (float)(MAX_IMAGINARY -\
-    (y) * ((MAX_IMAGINARY - minImaginary) / (float)((height) - 1)))
+    (double)(MAX_IMAGINARY -\
+    (y) * ((MAX_IMAGINARY - minImaginary) / (double)((height) - 1)))
 
 #define REAL_POS(x, width)\
-    (float)(minReal + (x) * ((maxReal - minReal) / (float)((width) - 1)))
+    (double)(minReal + (x) * ((maxReal - minReal) / (double)((width) - 1)))
 
 __kernel void calculate(__write_only image2d_t texture, const int2 size,
-        const int maxIterations, const float minReal, const float maxReal,
-        const float minImaginary)
+        const int maxIterations, const double minReal, const double maxReal,
+        const double minImaginary)
 {
     const int posX = get_global_id(0);
     const int posY = get_global_id(1);
@@ -24,10 +24,10 @@ __kernel void calculate(__write_only image2d_t texture, const int2 size,
     if (posX >= size.x || posY >= size.y)
         return;
 
-    float imaginaryNumber = IMAGINARY_POS(posY, get_image_height(texture));//getImaginary(posY, get_image_height(texture));
-    float realNumber = REAL_POS(posX, get_image_width(texture));//getReal(posX, get_image_width(texture));
-    float zReal, zImaginary, tmp;
-    float zRealSquared, zImaginarySquared;
+    double imaginaryNumber = IMAGINARY_POS(posY, get_image_height(texture));//getImaginary(posY, get_image_height(texture));
+    double realNumber = REAL_POS(posX, get_image_width(texture));//getReal(posX, get_image_width(texture));
+    double zReal, zImaginary, tmp;
+    double zRealSquared, zImaginarySquared;
     float color;
     bool isInside = true;
